requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
distro==1.9.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.3
google-oauth2-tool==0.0.3
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
httplib2==0.20.2
idna==3.10
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.4
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==3.0.2
more-itertools==8.10.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
oauth2client==4.1.3
oauthlib==3.2.0
packaging==24.2
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.11.2
pyparsing==2.4.7
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
soupsieve==2.6
sympy==1.13.1
tqdm==4.67.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
wadllib==1.3.6
zipp==1.0.0
